Refactored PuzzleManager to LevelManager
