+++ Assets/Scenes/_MAINSCENE_1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 74f598e239b39f04fb0e3ab5fc502fce
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
