+++ Assets/Scripts/Game/GameController.cs
@@ -8,12 +8,12 @@ public class GameController : MonoBehaviour
 {
   public static GameController Instance { get; private set; }
   public PlayerManager PlayerManager;
-  public PuzzleManager PuzzleManager;
+  public LevelManager levelManager;
   [SerializeField] public bool forceStart = false;
 
   void Awake()
   {
-    EventManager.AddListener<GameStartEvent>(onGameStart);
+    EventManager.AddListener<LevelStartEvent>(onGameStart);
 
     if (Instance != null && Instance != this)
     {
@@ -56,11 +56,11 @@ public class GameController : MonoBehaviour
   {
     return forceStart || PlayerManager.players.Count == 4;
   }
-  public void LoadPuzzle()
+  public void LoadLevel()
   {
     if (checkCanStartGame())
     {
-      LoadScene("_MAINSCENE");
+      LoadScene("_MAINSCENE_1");
     }
     else
     {
@@ -70,19 +70,19 @@ public class GameController : MonoBehaviour
 
   public bool IsGameEnding()
   {
-    if (PuzzleManager == null)
+    if (levelManager == null)
     {
       return false;
     }
     else
     {
-      return PuzzleManager.gameIsEnding;
+      return levelManager.gameIsEnding;
     }
   }
 
-  private void onGameStart(GameStartEvent e)
+  private void onGameStart(LevelStartEvent e)
   {
     Debug.Log("Game Start received by GameController");
-    PuzzleManager = GameObject.Find("PuzzleManager").GetComponent<PuzzleManager>();
+    levelManager = GameObject.Find("LevelManager").GetComponent<LevelManager>();
   }
 }
