+++ Assets/Scripts/Puzzles/Puzzle.cs
@@ -1,0 +1,30 @@
+﻿using System;
+using UnityEngine;
+
+public abstract class Puzzle : MonoBehaviour
+{
+    public LevelManager levelManager;
+    public delegate void completeCallback();
+    private completeCallback _complete_callback;
+    public bool isComplete = false;
+
+    public virtual void Awake()
+    {
+        levelManager = GameObject.Find("LevelManager").GetComponent<LevelManager>();
+        levelManager.addPuzzle(this);
+    }
+
+    public void SetCompleteCallback(completeCallback callback)
+    {
+        _complete_callback = callback;
+    }
+    
+    public void Complete()
+    {
+        isComplete = true;
+        if (_complete_callback != null)
+        {
+            _complete_callback();
+        }
+    }
+}
