+++ Assets/Scripts/Puzzles/Puzzle.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a532d9fdbc66458e8ee6f926c52828b7
+timeCreated: 1665538595
