+++ Assets/Scripts/Puzzles/TypeWritterPuzzle.cs
@@ -1,0 +1,152 @@
+﻿using UnityEngine;
+using System.Collections;
+using System.Collections.Generic;
+using TMPro;
+
+public enum ButtonType
+{
+    Submit,
+    A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z,
+    One, Two, Three, Four, Five, Six, Seven, Eight, Nine, Zero,
+    Space,
+}
+
+public class TypeWritterPuzzle : Puzzle
+{
+    private TextMeshProUGUI puzzle_text;
+
+    [Header("Puzzle")]
+    [Tooltip("This defines the possible puzzle solutions.")]
+    public static List<List<ButtonType>> Solutions = new List<List<ButtonType>> {
+        new List<ButtonType> {
+          ButtonType.B, ButtonType.L, ButtonType.U, ButtonType.E, ButtonType.Space,
+          ButtonType.R, ButtonType.E, ButtonType.D, ButtonType.Space,
+          ButtonType.Y, ButtonType.E, ButtonType.L, ButtonType.L, ButtonType.O, ButtonType.W
+          },
+      };
+    
+      private Dictionary<ButtonType, string> ButtonToString = new Dictionary<ButtonType, string> {
+        {ButtonType.A, "A"},
+        {ButtonType.B, "B"},
+        {ButtonType.C, "C"},
+        {ButtonType.D, "D"},
+        {ButtonType.E, "E"},
+        {ButtonType.F, "F"},
+        {ButtonType.G, "G"},
+        {ButtonType.H, "H"},
+        {ButtonType.I, "I"},
+        {ButtonType.J, "J"},
+        {ButtonType.K, "K"},
+        {ButtonType.L, "L"},
+        {ButtonType.M, "M"},
+        {ButtonType.N, "N"},
+        {ButtonType.O, "O"},
+        {ButtonType.P, "P"},
+        {ButtonType.Q, "Q"},
+        {ButtonType.R, "R"},
+        {ButtonType.S, "S"},
+        {ButtonType.T, "T"},
+        {ButtonType.U, "U"},
+        {ButtonType.V, "V"},
+        {ButtonType.W, "W"},
+        {ButtonType.X, "X"},
+        {ButtonType.Y, "Y"},
+        {ButtonType.Z, "Z"},
+        {ButtonType.One, "1"},
+        {ButtonType.Two, "2"},
+        {ButtonType.Three, "3"},
+        {ButtonType.Four, "4"},
+        {ButtonType.Five, "5"},
+        {ButtonType.Six, "6"},
+        {ButtonType.Seven, "7"},
+        {ButtonType.Eight, "8"},
+        {ButtonType.Nine, "9"},
+        {ButtonType.Zero, "0"},
+        {ButtonType.Space, " "},
+      };
+      
+      [SerializeField] private List<ButtonType> _pressed = new List<ButtonType> { };
+
+      [SerializeField] private List<ButtonType> _solution;
+      [SerializeField] private List<ButtonType> _answer = new List<ButtonType> { };
+      
+      public override void Awake()
+      {
+          base.Awake();
+          EventManager.AddListener<InteractEvent>(OnButtonPressed);
+      }
+      
+      void Start()
+      {
+          puzzle_text = GameObject.Find("Puzzle_Text").GetComponentInChildren<TextMeshProUGUI>();
+
+          // Demo puzzle ID = 0
+          UpdateSolution(0);
+          LevelStartEvent levelStartEvent = new LevelStartEvent();
+          EventManager.Broadcast(levelStartEvent);
+          Debug.Log("GameStartEvent broadcasted");
+      }
+      
+      void UpdateSolution(int puzzleId)
+      {
+          _solution = Solutions[puzzleId];
+      }
+      
+      public void OnButtonPressed(InteractEvent evt)
+      {
+          Button button = evt.gameObject.GetComponent<Button>();
+          if (button)
+          {
+              ButtonPressed(button);
+          }
+      }
+      
+      public void ButtonPressed(Button button)
+      {
+          _pressed.Add(button.buttonType);
+          if (_answer.Count == 0)
+          {
+              puzzle_text.text = "";
+          }
+          switch (button.buttonType)
+          {
+              case ButtonType.Submit:
+                  if (CheckAnswer())
+                  {
+                      levelManager.audioController.playSuccess();
+                      Complete();
+                  }
+                  else
+                  {
+                      _answer.Clear();
+                      puzzle_text.text = "Mistake!";
+                      levelManager.audioController.playMistake();
+                  }
+                  break;
+              default:
+                  _answer.Add(button.buttonType);
+                  puzzle_text.text += ButtonToString[button.buttonType];
+                  break;
+          }
+      }
+      
+      private bool CheckAnswer()
+      {
+          if (_answer.Count == _solution.Count)
+          {
+              for (int i = 0; i < _answer.Count; i++)
+              {
+                  if (_answer[i] != _solution[i])
+                  {
+                      return false;
+                  }
+              }
+
+              return true;
+          }
+          else
+          {
+              return false;
+          }
+      }
+}
