+++ Assets/Scripts/Puzzles/TypeWritterPuzzle.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1120b812f108429dae8f4a0c8b567643
+timeCreated: 1665538645
