+++ ProjectSettings/EditorBuildSettings.asset
@@ -14,7 +14,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/End.unity
     guid: 34b10751cfaa4a840baa0dbd96eb500e
-  - enabled: 0
-    path: Assets/Scenes/MP_MAINSCENE.unity
-    guid: cb3afe1b486b1744b8299395fce0c8dc
+  - enabled: 1
+    path: Assets/Scenes/_MAINSCENE_1.unity
+    guid: 74f598e239b39f04fb0e3ab5fc502fce
   m_configObjects: {}
